feat: Converted example to /app directory
