--- examples/nextjs-yjs-slate/app/globals.css
@@ -1,41 +1,0 @@
-html,
-body {
-  padding: 0;
-  margin: 0;
-  font-family: -apple-system, BlinkMacSystemFont, Segoe UI, Roboto, Oxygen,
-  Ubuntu, Cantarell, Fira Sans, Droid Sans, Helvetica Neue, sans-serif;
-}
-
-a {
-  color: inherit;
-  text-decoration: none;
-}
-
-* {
-  box-sizing: border-box;
-}
-
-main {
-  position: fixed;
-  inset: 0;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  padding: 16px;
-  background: #e8eaee;
-}
-
-.loading {
-  position: absolute;
-  width: 100vw;
-  height: 100vh;
-  display: flex;
-  place-content: center;
-  place-items: center;
-}
-
-.loading img {
-  width: 64px;
-  height: 64px;
-  opacity: 0.2;
-}
