+++ examples/nextjs-yjs-slate/next.config.js
@@ -1,5 +1,8 @@
-/** @type {import('next').NextConfig} */
+/** @type {import("next").NextConfig} */
 const nextConfig = {
+  experimental: {
+    appDir: true,
+  },
   reactStrictMode: true,
   transpilePackages: ["@slate-yjs/react"],
 };
