--- examples/nextjs-yjs-slate/pages/_app.tsx
@@ -1,6 +1,0 @@
-import "@/styles/globals.css";
-import type { AppProps } from "next/app";
-
-export default function App({ Component, pageProps }: AppProps) {
-  return <Component {...pageProps} />;
-}
