--- examples/nextjs-yjs-slate/pages/api/auth.ts
@@ -1,64 +1,0 @@
-import { authorize } from "@liveblocks/node";
-import { NextApiRequest, NextApiResponse } from "next";
-
-const API_KEY = process.env.LIVEBLOCKS_SECRET_KEY;
-
-export default async function auth(req: NextApiRequest, res: NextApiResponse) {
-  if (!API_KEY) {
-    return res.status(403).end();
-  }
-
-  const room = req.body.room;
-
-  const response = await authorize({
-    room,
-    secret: API_KEY,
-    userId: `user-${Math.floor(Math.random() * 1000)}`,
-    userInfo: USER_INFO[Math.floor(Math.random() * 10) % USER_INFO.length],
-  });
-
-  return res.status(response.status).end(response.body);
-}
-
-export const USER_INFO = [
-  {
-    name: "Charlie Layne",
-    color: "#D583F0",
-    picture: "https://liveblocks.io/avatars/avatar-1.png",
-  },
-  {
-    name: "Mislav Abha",
-    color: "#F08385",
-    picture: "https://liveblocks.io/avatars/avatar-2.png",
-  },
-  {
-    name: "Tatum Paolo",
-    color: "#F0D885",
-    picture: "https://liveblocks.io/avatars/avatar-3.png",
-  },
-  {
-    name: "Anjali Wanda",
-    color: "#85EED6",
-    picture: "https://liveblocks.io/avatars/avatar-4.png",
-  },
-  {
-    name: "Jody Hekla",
-    color: "#85BBF0",
-    picture: "https://liveblocks.io/avatars/avatar-5.png",
-  },
-  {
-    name: "Emil Joyce",
-    color: "#8594F0",
-    picture: "https://liveblocks.io/avatars/avatar-6.png",
-  },
-  {
-    name: "Jory Quispe",
-    color: "#85DBF0",
-    picture: "https://liveblocks.io/avatars/avatar-7.png",
-  },
-  {
-    name: "Quinn Elton",
-    color: "#87EE85",
-    picture: "https://liveblocks.io/avatars/avatar-8.png",
-  },
-];
