+++ examples/nextjs-yjs-slate/src/app/layout.tsx
@@ -1,4 +1,4 @@
-import "./globals.css";
+import "@/globals.css";
 
 export const metadata = {
   title: "Liveblocks",
