+++ examples/nextjs-yjs-slate/src/components/Avatars.module.css
