+++ examples/nextjs-yjs-slate/src/components/Loading.tsx
