+++ examples/nextjs-yjs-slate/src/globals.css
@@ -3,7 +3,7 @@ body {
   padding: 0;
   margin: 0;
   font-family: -apple-system, BlinkMacSystemFont, Segoe UI, Roboto, Oxygen,
-    Ubuntu, Cantarell, Fira Sans, Droid Sans, Helvetica Neue, sans-serif;
+  Ubuntu, Cantarell, Fira Sans, Droid Sans, Helvetica Neue, sans-serif;
 }
 
 a {
@@ -23,19 +23,4 @@ main {
   align-items: center;
   padding: 16px;
   background: #e8eaee;
-}
-
-.loading {
-  position: absolute;
-  width: 100vw;
-  height: 100vh;
-  display: flex;
-  place-content: center;
-  place-items: center;
-}
-
-.loading img {
-  width: 64px;
-  height: 64px;
-  opacity: 0.2;
 }
